Initial clean up
